Rename `mockData` to `mock_data`
