+++ lib/views/course_view/downloaded_courses_view/downloaded_courses_view.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:gocast_mobile/models/user/mockData.dart';
+import 'package:gocast_mobile/models/user/mock_data.dart';
 import 'package:gocast_mobile/providers.dart';
 import 'package:gocast_mobile/views/components/base_view.dart';
 import 'package:gocast_mobile/views/course_view/downloaded_courses_view/content_view.dart';
